requirements.txt
pandas
matplotlib
plotly
sqlalchemy
openpyxl
psycopg2-binary
